requirements.txt
streamlit
pandas
python-docx
openpyxl
tqdm
PyPDF2